Update location and frame center updaters to handle GeometryCollections
The geometry union operations failed on some test videos because we were
using the union(Geometry) function, which does not support
GeometryCollections. It is possible for the consecutive unioning of
child metacard geometries to result in a GeometryCollection because some
child metacards (chunks) may have only a single Point and others may
have LineStrings, for example, and unioning them results in a
GeometryCollection that has to be unioned with the remaining child
geometries.

The updated method of unioning the geometries is also the preferred way
to do it, according to GeoTools:
https://docs.geotools.org/latest/userguide/library/jts/combine.html